Fix for issue https://github.com/epics-modules/iocStats/issues/20
As indicated by the MinGW group the bug is actually the missing definition of WIN32 version, not with the usage of the function itself since it is only available in Windows 7+. Since most computers should be running Windows 7 by now it probably can be assumed to be the case that this should be available. Epics Base has already implemented the fix in their repo to link the psapi library so the only leftover change to implement is this.